+++ diamondsymbol.h
@@ -16,8 +16,8 @@ protected:
   void addDiamond(QPainterPath& path, const QRectF& rect);
 
 private:
-  int m_w;
-  int m_h;
+  qreal m_w;
+  qreal m_h;
 };
 
 #endif /* __DIAMONDSYMBOL_H__ */
